beam_bsm: Improve handling of bs_start_match2 instructions
The handling of bs_start_match2 was both too conservative and too
careless. It was too conservative in that would not do the
optimization if the were copies of the match state in other
registers. It was careless in that it did not consider the
failure branch.

Reorganize the code and fix both these issues. Add a test case
to test that the failure branch is considered.