Refactor code generation for decoding SEQUENCE/SET
In order to optimize the generated code, we will need data structures
that can be traversed in several passes before the actual code is
generated.

Refactor the code so that if first builds (essentially) a list of
funs that do the actual code generation, and a second function
that will call the funs one at the time.