erts: Change default of erl_crash.dump
* Create an erl_crash.dump if no heart exists
  and no ERL_CRASH_DUMP_SECONDS is set (behaviour changed)
* Don't create an erl_crash.dump if heart do exists
  and no ERL_CRASH_DUMP_SECONDS is set (behaviour not changed)

This changes the behaviour back to the R15B02 default considering
if a beam was running with no heart.

The motivation is that we want a fast heart to fast restart erlang if
it crashes but to be able to get non-truncated erl_crash.dumps via
setting ERL_CRASH_DUMP_SECONDS if we so desire.