erts,hipe: Make hipe compiler ask running emulator about process struct offsets
Changes in "struct process" has made native code more depenedent on if the
executing emulator is smp enabled or not.

This commit will make a default built hipe compiler (without -xcomp)
to ask the running emulator about process struct offsets. This will
make native code work (again) as long as the same emulator is used for
compilation as well as execution of the native code.