[common_test] Stop cover on slave node after node is terminated
Before terminating slave nodes, common_test calls cover:flush/1 to
fetch data from the node without actually stopping cover on this
node. If cover is not stopped for the node and a new node with the
same name is started, then cover will be started on the new node. To
avoid this common_test now calls cover:stop/1 after the slave node is
terminated.