escript to accept emulator arguments when script file has no shebang
According to the documentation, if the second or third line in a
script starts with %%!, then escript will use the rest of the line
as emulator options.  However, previously this was only the case
if the first line started with #!.  This change removes that check,
and unconditionally uses the %%! line if present.