Make trace_local_SUITE work without the reverse eta conversion
The exceptions tracing tests look for {trace_local_SUITE,exc,2} in
stacktraces to strip them and compare them to the excepted result, the
removal of the reverse eta conversion renames '-exception_test/3-fun-0-'
to exc and thus confuses the stripping code. This commit fix this by
not using `fun exc/2` but `fun (F, As) -> exc(F, As) end` instead, which
was what the reverse eta conversion was doing.