Extend char() to Unicode characters
The code related to the introduction of unicode_string() and
unicode_char() has been removed. The types char() and string() have
been extended to include Unicode characters.

In fact char() was changed some time ago; this commit is about
cleaning up the documentation and introduce better names for some
functions.