Remove one use of iolist_size/1 in io_lib_pretty.erl
In order to simplify the testing of Unicode atoms a minor adjustment
has been made in io_lib_pretty. Unicode atoms will not be introduced
until R18 but it may come in handy to be able to pretty-print such
atoms in R16. This update affects ~p only; ~w already works.