erts: Fix ASSERT crash in erts_try_seize_code_write_permission
ASSERT(code_writing_process != c_p);

The assert was actually harmless and could be removed as one solution.
But I think it's better to keep it as a way to catch bugs that would
otherwise lead to horrible deadlocks if a thread "forgets" to call
erts_release_code_write_permission().

As a result erts_release_code_write_permission() must be called BEFORE
resuming the suspended process.