Remove error handling for calling a BIF that is not auto-imported
All guards BIFs are auto-imported. That can be verified like this:

  [] = [{F,A} || {erlang,F,A} <- erlang:system_info(snifs),
                                 erl_internal:guard_bif(F, A),
                                 not erl_internal:bif(F, A)]

Therefore, calling a guard BIF in a guard without a module name
is always allowed (provided that there is not a local function or
an import with the same name), and therefore we can remove the
error reporting code. But keep an assertion so that we will find
out if any non-auto-imported guard BIFs are added in the future.