Make prim_file skip invalid filenames in unicode mode
The fix affects list_dir and read_link. Raw filenames are now
never produced, just consumed even if +fnu or +fna is used on
Linux etc.

This also adds the options to get error return or error handler
warning messages with +fn{u|a}{i|w|e} as an option to erl.

This is still not documented and there needs to be other versions
of read_dir and read_link to facilitate reading of all types
of filenames and links.

A check that we will not change to an invalid directory is also needed.