Forbid multiple values in Core Erlang sequence arguments
It does not make sense to return multiple values from a sequence
argument and the Kernel Erlang passes can't cope with it.

The linting pass now knows how to detect this kind of defunct code and
the Core code folding pass is changed to not generate code like that
when optimizing away multiple-valued lets in effect mode.

Reported-by: José Valim