Clean up of documentation of type language
Prompted by noticing that the type language description was a bit vague
in its explanation of the syntax and semantics of bitstring types,
I decided to add this information.  While at it, a general clean up and
rephrasing was performed.  The binary() type is now described as an alias,
as it should be, and a bitstring() type is also described as a built-in
alias.