[reltool] Make escript test work with hipe and smp
reltool_server_SUITE:create_standalone and create_multiple_standalone
failed on test hosts running hipe and smp. The reason was that the
escript used in these tests explicitly disabled smp, which in turn
made the emulator spit out warnings like this:

"<HiPE (v 3.10)> Warning: not loading native code for module beam_lib:
it was compiled for an incompatible runtime system; please regenerate
native code for this runtime system"

This was returned from the escript and did not match the expected result.

To overcome this, the escript (reltool/examples/display_args) is
changed and does no longer disable smp.