Extend cover.erl source search
Commit 29231033 made cover fall back to compile info if source was not
found in pwd or ../src. This isn't sufficient for source that lies in
subdirectories of ../src when beams and source have been moved since
compilation (eg. install of some OTP applications), so first try finding
source relative to the beam directory.

For example, given a beam path

  /installed/path/to/app-1.0/ebin/root.beam

and a source path

  /compiled/path/to/app/src/subdir/root.erl

then look for

  /installed/path/to/app-1.0/ebin/../src/subdir/root.erl

before the source path.