asn1_constructed_per: Optimize decoding of CHOICE
Optimize the decoding of CHOICE. Most important is to inline decoding
of the extension bit (if present) and decoding of the choice index
to give the BEAM compiler more opportunities for optimization.

We will also change the structure of the generated code. The current
code uses a flattened case for both the root and extension alternatives:

  case Choice + NumRootChoices * Ext of
     %% Root alternatives.
     0 - ...;
     :
     LastRootAlternative -> ...;
     %% Extension alternatives.
     LastRootAlternative+1 -> ...;
     :
     %% Unknown extension.
     _ -> ...;
  end

We will instead generate nested cases:

  case Ext of
    0 ->
      case Choice of
        %% Root alternatives.
        0 - ...;
        :
        LastRootAlternative -> ...
       end;
    1 ->
       %% Extension alternatives.
       <Decode the open type here>
       case Choice of
         0 -> ...;
         :
	 LastExtensionAlternative -> ...;
         %% Unknown extension.
         _ -> ...;
       end
    end

Nested cases should be slightly faster. For decoding of the extensions,
it also makes it possible to hoist the decoding of the open type up
from each case to before the case switching on the extension index,
thus reducing the size of the generated code.

We will also do another change to the structure. Currently, the
big flat clase is wrapped in code that repackages the return values:

  {Alt,{Value,RemainingEncodedData}} =
     case Choice + NumRootChoices * Ext of
       :
     end,
  {{Value,Alt},RemainingEncodedData}.

We still need to do the repackaging, but we can push it down to
the case arm for decoding each alternative. In many cases, that
will give the BEAM compiler the opportunity to avoid building the
temporary tuples.