asn1ct_name: Simplify the data structures of the name server process
Since we no longer need to support push/1 and pop/1, we can simplify
the data structures used for keeping track of the variables.

Each entry in the list can be simply {Var,integer()} instead of
{Var,[integer()]}.

While at it, eliminate calls to the obsolete lists:keysearch/3
function, don't use integer_to_list/0 within a call to lists:concat/1,
and the catch all clause in get_prev/2 and get_next/2.