[debugger] Add an option 'Strings'
A new checkbox has been added. When it is checked, the range set by
the 'erl' flag '+pc' is used for determining when to print lists of
integers as strings. When it is unchecked, integer lists are never
printed as strings.

A minor incompatibility: settings saved by Erlang R16B01
or later cannot be read by Erlang R16B or earlier.