Fix bug related to the handling of is_number/1 by the range analysis
The following module produced erroneous results when compiled with HiPE:

-module(a).
-export([foo/1]).

foo(X) when is_number(X) ->
    is_integer(X).

Running:

  1> c(a).
  2> a:foo(0).
  true
  3> hipe:c(a).
  4> a:foo(0).
  false      % *** WRONG ***

The problem was that the 'number' case for the `hipe_icode:type_test/1`
was going to the default case where the argument was determined as being
something other than an integer. Thanks to Sebastian Egner and Johannes
Weißl for bringing the bug into attention. Fixed on the day it was reported.