asn1_test_lib: Give more information when things goes wrong
Ensure that compilation errors of Erlang code gets printed.

If compilation of an ASN.1 specification goes wrong, print out
the filename of the offending specification. It can be seen
in the test case log, but because most test cases run in a
parallel group, the test case log is not available until all
test case in the group have finished.