Fix handling of bs_match_string translation
Prior to R16B, the bs_match_string instruction worked only for cases
that had a destination variable. As shown by the example below, this
is not always the case and the Dst list can be empty.

  test(<<42, _/bits>> = B) -> B.

This fixes a bug reported by Loïc Hoguin (02/12/2013); the original
post used a different code example to illustrate the problem.

While revising the code, a significant clean up was also performed.