Eliminate latin1 encoded characters
Characters encoded in latin1 (with no encoding comment) will not
work in R17 since the default encoding will be UTF-8.

The asn1 test suite contains two instances of comments written in
Swedish.

The comment in the .asn1config file is hard to understand and possibly
out-of-date. Since the comment is not essential, simply delete it.

Translate the comment in the other file to English.