add sys:get_state/1,2 and sys:replace_state/2,3
At Erlang Factory 2013 there was discussion during one of the talks about
the sys:get_status functions and how useful they were for debugging. Geoff
Cant mentioned it would be very useful if the sys module also provided
functions to use while debugging to get just the state of a process and
also to be able to replace the state of a process, and many others in the
audience appeared to agree.

The sys:get_state/1,2 functions return the state of a gen_server, gen_fsm,
or gen_event process. The return value varies depending on the process
type: process state for a gen_server, state name and state data for a
gen_fsm, and handler module, handler id, and handler state for each handler
registered in a gen_event process.

The sys:replace_state/2,3 functions allow the state of a gen_server,
gen_fsm, or gen_event process to be replaced with a new state. These
functions take a function argument that updates or replaces the process
state; using a function to change the state eliminates the race condition
of first reading the state via sys:get_state/1 or sys:get_state/2, using
its return value to create a new state, and then replacing the old state
with the new state, since during that time the process might have received
other calls or messages that could have changed its state.

* For a gen_server process, the state replacement function takes the
  process state as an argument and returns a new state.

* For a gen_fsm process, the state replacement function gets a tuple of
  {StateName, StateData} and returns a similar tuple that specifies a new
  state name, new state data, or both.

* For a gen_event process, the state replacement function is called for
  each registered event handler. It gets a tuple {Module, Id, HandlerState}
  and returns a similar tuple that specifies the same Module and Id values
  but may specify a different value for HandlerState.

If the state replacement function crashes or results in an error, the
original state of a gen_server or gen_fsm process is maintained; if such a
crash occurs for a gen_event process, the original state of the event
handler for which the state replacement function was called is maintained,
but the states of other event handlers of the same gen_event process may
still be updated if no errors or crashes occur while replacing their
states.

Add documentation for sys:get_state/1,2 and sys:replace_state/2,3. The
documentation explicitly notes that the functions are intended for use
during debugging.

Add new tests for these functions to gen_server_SUITE, gen_fsm_SUITE, and
gen_event_SUITE.