Fix rest_for_one and one_for_all restarting a child not terminated
In rest_for_one and one_for_all supervisors one child dying can cause
multiple children to be restarted. Previously if the child that caused
the restart is started successfully but another child fails to start,
the supervisor would not terminate this child with the other
successfully restarted children as no record of the pid was kept. Thus
the supervisor would try to start this child again. This could lead to
multiples of the same child or if the child is registered cause repeated
attempts at starting this child - until the max restart threshold was
reached.

Now the child that failed to start becomes the restarting child, instead
of staying with the same child, for the next restart attempt. This has
the following side effects:

1) In one_for_all the new version of the child that original died is
   terminated before a restart attempt is made.

2) In rest_for_one all succesfully restarted children are not terminated
   and restarting continues from the child that failed to start.